pyproject.toml
[tool.poetry]
name = "rpg-bot-monorepo"
version = "0.1.0"
description = "Monorepo for RPG-Bot: contracts, services, memory and generation packages."
authors = ["RPG-Bot Team <dev@rpg-bot.example>"]
license = "Proprietary"
readme = "README.md"
packages = [
    { include = "rpg_gateway_api", from = "services/gateway_api/src" },
    { include = "rpg_party_sync", from = "services/party_sync/src" },
    { include = "rpg_media_broker", from = "services/media_broker/src" },
    { include = "memory37", from = "packages/memory37/src" },
    { include = "genlayers", from = "packages/genlayers/src" },
    { include = "rpg_contracts", from = "packages/rpg_contracts/src" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = { version = "^0.30.1", extras = ["standard"] }
httpx = "^0.27.0"
pydantic = "^2.7.4"
pydantic-settings = "^2.3.4"
pyjwt = "^2.8.0"
python-dateutil = "^2.9.0.post0"
pyyaml = "^6.0.1"
python-json-logger = "^2.0.7"
opentelemetry-sdk = "^1.28.2"
opentelemetry-exporter-otlp = "^1.28.2"
opentelemetry-instrumentation-fastapi = "^0.49b2"
opentelemetry-instrumentation-requests = "^0.49b2"
prometheus-fastapi-instrumentator = "^7.0.0"
openai = "^1.0.0"
psycopg = { version = "^3.2.0", extras = ["binary"] }
jsonschema = "^4.23.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-asyncio = "^0.23.6"
pytest-cov = "^5.0.0"
ruff = "^0.6.1"
mypy = "^1.10.0"
types-pyyaml = "^6.0.12.20240808"
types-python-dateutil = "^2.9.0.20240316"
jsonschema = "^4.23.0"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q --disable-warnings"
testpaths = ["services", "packages", "qa/tests"]
asyncio_mode = "strict"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ASYNC", "B", "BLE", "C4", "DTZ", "EXE", "G", "PIE", "PT", "RET", "RUF", "SIM"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_unreachable = true
plugins = ["pydantic.mypy"]
show_error_codes = true

[tool.coverage.run]
branch = true
source = [
    "services/gateway_api/src",
    "services/party_sync/src",
    "services/media_broker/src",
    "packages/memory37/src",
    "packages/genlayers/src",
    "packages/rpg_contracts/src"
]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
